+++ backend/dataall/modules/worksheets/__init__.py
@@ -26,7 +26,7 @@ class WorksheetApiModuleInterface(ModuleInterface):
 
     @staticmethod
     def depends_on() -> List[Type['ModuleInterface']]:
-        from dataall.modules.datasets import DatasetApiModuleInterface
+        from dataall.modules.s3_datasets import DatasetApiModuleInterface
 
         return [DatasetApiModuleInterface]
 
@@ -45,6 +45,6 @@ class WorksheetCdkModuleInterface(ModuleInterface):
 
     @staticmethod
     def depends_on() -> List[Type['ModuleInterface']]:
-        from dataall.modules.datasets import DatasetCdkModuleInterface
+        from dataall.modules.s3_datasets import DatasetCdkModuleInterface
 
         return [DatasetCdkModuleInterface]
