+++ frontend/src/modules/Datasets/views/DatasetCreateForm.js
@@ -51,10 +51,10 @@ const DatasetCreateForm = (props) => {
   const [groupOptions, setGroupOptions] = useState([]);
   const [environmentOptions, setEnvironmentOptions] = useState([]);
   const [confidentialityOptions] = useState(
-    config.modules.datasets.features.confidentiality_dropdown === true &&
-      config.modules.datasets.features.custom_confidentiality_mapping
+    config.modules.s3_datasets.features.confidentiality_dropdown === true &&
+      config.modules.s3_datasets.features.custom_confidentiality_mapping
       ? Object.keys(
-          config.modules.datasets.features.custom_confidentiality_mapping
+          config.modules.s3_datasets.features.custom_confidentiality_mapping
         )
       : ConfidentialityList
   );
@@ -396,7 +396,7 @@ const DatasetCreateForm = (props) => {
                           </Box>
                         </CardContent>
                         <CardContent>
-                          {config.modules.datasets.features
+                          {config.modules.s3_datasets.features
                             .auto_approval_for_confidentiality_level[
                             values.confidentiality
                           ] === true && (
